Added Circular Priority Queue.
The gateway supports messages with timestamp (ts).
If some package are out of order a priority queue will order them.
The priority queue is based on a min heap.

If a publish message does not have timestamp, the timestamp will be
added by the gateway.

The samples were updates accordingly.
The ESP8266 sensor sample uses a NTP client library to get the correct
time.

The websites use the ts field in the json instead of the system time.
A third website sample was developed (sensor.html).
It is based on the previous BME280, but with and added gauge.
The javacript portions of the samples were moved to separated files.
A default css file was created.

Some minor tweaks are missing.